create models for user and tour
