+++ travxy/resources/user.py
@@ -1,0 +1,28 @@
+import sqlite3 
+from flask_restful import Resource, reqparse
+from travxy.models.user import UserModel
+
+class UserRegister(Resource):
+    parser = reqparse.RequestParser()
+    parser.add_argument('username',
+                        type=str, 
+                        required=True, 
+                        help="This Field cannot be blank")
+
+    parser.add_argument('password',
+                        type=str, 
+                        required=True, 
+                        help="This Field cannot be blank")
+    def post(self):
+        data = UserRegister.parser.parse_args()
+        if UserModel.find_by_username(data['username']):
+            return {"message":"A user with that username already exist"}, 400
+
+        connection = sqlite3.connect('trav_data.db')
+        cursor = connection.cursor()
+        query = "INSERT INTO users VALUES(NULL, ?, ?)"
+        cursor.execute(query, (data['username'], data['password']))
+
+        connection.commit()
+        connection.close()
+        return {"message": "User created succesfully"}, 201
