+++ travxy/security.py
@@ -1,13 +1,13 @@
 from hmac import compare_digest
-from travxy.user import User
+from travxy.models.user import UserModel
 
 from werkzeug.security import safe_str_cmp
 
 def authenticate(username, password):
-    user = User.find_by_username(username)
+    user = UserModel.find_by_username(username)
     if user and compare_digest(user.password, password):
         return user
 
 def identity(payload):
     user_id = payload['identity']
-    return User.find_by_id(user_id)
+    return UserModel.find_by_id(user_id)
